Fix `validateBigIntegerScale()` to catch `Integer.MIN_VALUE` scale
`Math.abs(Integer.MIN_VALUE)` is itself negative, so that one scale slipped
past the magnitude check and reached `BigDecimal.toBigInteger()` -- the #968
expansion the guard exists to prevent. Widen to `long`.

Not reachable from JSON parsing (exponent overflow is rejected earlier), but
`validateBigIntegerScale()` is public API callable with any `BigDecimal`.

+++ src/main/java/tools/jackson/core/StreamReadConstraints.java
@@ -654,7 +654,9 @@ public class StreamReadConstraints
      */
     public void validateBigIntegerScale(int scale) throws StreamConstraintsException
     {
-        final int absScale = Math.abs(scale);
+        // 15-Sep-2026, tatu: [core#1715] Widen to `long`: `Math.abs(Integer.MIN_VALUE)`
+        //    is itself negative, and would slip past the check below
+        final long absScale = Math.abs((long) scale);
         final int limit = getMaxBigIntegerScale();
 
         if (absScale > limit) {
